Client: Make sure that fast restarts are not looped
This adds a mechanism that makes sure that restarts that take a small
amount of time `delta` are not looped. See the Doc string for
restartUntilErr for details.

The solves the issue that if an error occurs and it is returned almost
immediately, the server will not be spammed with requests because we
get stuck in a restart loop. Instead of also the hardcoded 2 second
wait, we do a 1, 2, 4, 8 second cycle which is also reset if the
function itself already takes longer than `delta`.